Move the EigenFilters header to Util since it's useful.
